requirements.txt
Django==3.9
djangorestframework==3.12.4
django-allauth
django-grappelli
Pillow
djangorestframework-simplejwt